Infrastructure for a neutron request handler
Adding a neutron handler. This is only the generic infrastructure
to handle all requests.
Handling of the specific requests will come in the next patch

Please note that most of this could be replaced by a off-the-shelf
REST framework later on. As this is not a priority, this has been postponed
to some later patch.

Change-Id: Ied923362bd935c079bfe1bc42000a8f4cd4a5492
Signed-off-by: Marcin Mirecki<mmirecki@redhat.com>