Connection to ovs db
This patch implements the connection to the OVS DB.
The OVS DB API lack a blocking connect method, so we
must implement this ourselves.
The implementation simply makes multiple retries to
the db, the pseudocode for it being:

  connection = ovs.db.idl.Idl
  while True:
    connection.run()
    if connection.has_ever_connected():
       return # Connection successful
    sleep()

A blocking connect method in OVS is not likely to
arrive soon. If it does, the code here should be
replaced by it.

Change-Id: Id4af4a723635182d24acd1217756e3456dbcacd0
Signed-off-by: Marcin Mirecki<mmirecki@redhat.com>