Avoid potential naming conflict by additional ovirt north db entries.
The provider keeps some additional information in north db.
This is needed to map the neutron API to the north db data format.
To avoid any conflict with any future OVN data, all this data must
be moved to the 'external_ids' columns. Additionaly all the extra
properties start with 'ovirt_'.

Change-Id: Idff0077369443c56650b263cc9a4f83f9bb21533
Signed-off-by: Marcin Mirecki<mmirecki@redhat.com>