spec: spell out %{python_sitelib}
For a reason unknown to me, el7's build-artifacts fails to substitute
%{python_sitelib}. Since %{python_sitelib} is anyway spelled out in
`make install`, we use the explicit path in the spec as well.

It would be nicer to use %{python_sitelib} in both cases, and make it
available in build-artifacts.el7, but let us unbreak the nightly builds
first.

Change-Id: Ifa2e8bcd9474da0edc1300b0ebb086ac5407de93
Signed-off-by: Dan Kenigsberg <danken@redhat.com>