Validate incoming Port devicd_id is not null or empty
Populating the name of the port is only temporary, as we set the name
to the port_id right after the first row commit. We set it to device_id
to avoid any duplication problems if parallel port creation should
happen.
Setting it to an empty value causes OVN to crash, so we must make sure
that the value is not empty.

Change-Id: I4e36f79ed101ae3d10ee2eb14baca16344a9f465
Signed-off-by: Marcin Mirecki <mmirecki@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1405094