BaseHandler sends errors as JSON
Without any configuration, the BaseHTTPRequestHandler
sends errors in HTML format to the client.
The BaseHandler configures the BaseHTTPRequestHandler
to format errors as JSON instead of HTML. This way all
responses formatted as JSON.
The structure of the error message is similar to OpenStack
Keystone's one, since there is no generic one.

The static error message of the BaseHTTPRequestHandler is
set to an empty string and an individual error message is
produced by BaseHandler.

Change-Id: I37b0f557328bcb64e1d25fba63452656f55c4e5c
Signed-off-by: Dominik Holler <dholler@redhat.com>