Config: Use default if a particular value is missing
Use default value if a particular configurable is missing in the
configuration file.
Before, the default value was used only if the whole configuration
file was not loaded. Now the default value is used if the
configuration file is loaded, too.

This is useful to ensure backward compatibility for the
configuration file on updates, which introduces new values in the
configuration file, e.g. SSL for HTTPS.

Change-Id: I8d5fcfd48c8afeb9a92014c0c9011fad4419a640
Signed-off-by: Dominik Holler <dholler@redhat.com>