Introduce Error Method Not Allowed (405)
Introduce the "Method Not Allowed" (http status 405) error to report
that the requested method is not allowed for the requested
resource, instead of reporting a Not Found (404) error.

This error is not strictly required for functionality, but improves
the error handling. According to RFC 7231 an "Allow" header field to
indicate the supported methods of resource is required, which is not
implemented to keep efforts low.

Change-Id: I1c5563333028134d0b96a0db2f3613d42cf137df
Signed-off-by: Dominik Holler <dholler@redhat.com>