Use time out in communication to engine
The auth.ovirt plugin use a connect and the read timeout
for the communication with engine. If such a timeout occurs,
this is reported as "Gateway Time-out" (http status 504), including
a detailed error message, to the client of the ovirt-provider-ovn.

This helps to distinguish between problems caused by the provider
and problems caused by engine.

Change-Id: Ie718ce64b26aa39d37e1178d59ab083f5ffcc112
Signed-off-by: Dominik Holler <dholler@redhat.com>