Define MTU explicitly
Define the MTU, which is set by DHCP to the VMs, explicitly.

Since OVN data is transported over an other network, the MTU
in the network provided by OVN should be small enough to
encapsulate one data unit of the tunneled data in one unit of
the tunneling protocol.

The default value for the MTU in the tunneling network is 1500,
defined by engine. The default MTU of the tunneled data is
determined by a dump of experimental network traffic, because the
used geneve tunneling protocol may contain optional fields.

Change-Id: Ie8a155e0995b48171bffaa7e96e3ed8fc55e0382
Bug-Url: https://bugzilla.redhat.com/1451342
Signed-off-by: Dominik Holler <dholler@redhat.com>