Update propagation of SSO error messages
Even if engine's SSO responds with HTTP error, the HTTP
message body data is parsed as json and the error message
is propagated by an exception.

This update is required, because engine's SSO changed in
f309fbfaec6f4df22e032161ee55e5c7fb5000ac the response
status for OAuthException to 400.

Change-Id: I70564f5062ac6e14bd6b44e43d6407865d14ac7c
Signed-off-by: Dominik Holler <dholler@redhat.com>