Modify spec Require's to ensure ovs beta package 1:2.6.1 is not consumed
This patch ensures that ovs package 1:2.6.1 is not
consumed when the provider is installed.
Ovs package 1:2.6.1 is a beta package which should
not be used.
When ovs is not installed, the result will be that the
correct version (if available) will be installed.
If 1:2.6.1 is already installed, the install operation
will fail with the message:
"... conflicts with openvswitch = 1:2.6.1 ..."
and manual action will be needed to replace the package.

The patch also updates the minimum required ovs version
to 2.7
Switch to ovs 2.8 will be made in another patch.

Change-Id: I634ef82f09c45e252687b88bb736ee74a9ad1c10
Signed-off-by: Marcin Mirecki <mmirecki@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1505398