Extract validation to separate file
The validation logic should be extracted to a separate file.
This patch creates a validation.py file which is to
hold all the validation logic.
It also moves the first bit of code (validating if
subnet has a gateway, required if a network is to be
attached to a router by subnet) to the new file.

Change-Id: I77527df39a64cc621e6ca4bd92a5c76212ab134d
Signed-off-by: Marcin Mirecki <mmirecki@redhat.com>