Store VLAN id as string
Currently the vlan id is not stored in external_ids, because
external_ids seems to store strings and discard integers silently.

Vlaned Physical networks are reported currently as flat networks,
because the vlan id is not stored as expected.

With this change vlaned Physical networks are reported as vlan
networks.

Change-Id: Ia043404e4bb300c57c344d6ba22178bd6d2833cb
Signed-off-by: Dominik Holler <dholler@redhat.com>