Implement allocation_pools for subnets
The information about a subnet in OpenStack Networking API is
extended with the 'allocation_pools', which are created from
the subnet's cidr. OVN skipps the first and last IPv4 address
of subnets, e.g. x.x.x.0 and x.x.x.255, and the second IPv4
address, e.g. x.x.x.1, is reserved for a router.

This change improves compatibility with ansible.
A test is added in change
Icf3438a102ea35128a48bb8a2e817455b808ce8e to OST.

Change-Id: Iac19d04104eaef560f56468d04d97a5465b73bde
Signed-off-by: Dominik Holler <dholler@redhat.com>