Redirect uncaught exceptions to provider log file
Uncaught exceptions in spawned threads are normally logged
so syslog. This can occur in the ovsdbapp spawned thread.
This patch adds a hook that will print the error message
into the provider log.

The patch also includes a workaround for a python
sys.excepthook thread bug:
http://bugs.python.org/issue1230540
which must be used to allow using the excepthook
with threading.

Change-Id: Idcc362a76adedeeea7f0c39b6de3fb5778821abe
Signed-off-by: Marcin Mirecki <mmirecki@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1537084