after_get_caps vdsm hook: report OpenStack binding:host_id
To satisfy OpenStack Networking API engine needs to know how
the host-id of the host. This new hook reports it in
getCapabilities. With this patch, vdsm now reports a
'openstack_binding_host_id' dictionary, mapping a provider
plugin type to host-id used by the plugin.
This suggested API allows for future assignment of a host
to more than a single external network provider.

Change-Id: Ic50537f1cf39577dcb96e104d20c2fe55a5d5d21
Signed-off-by: Dominik Holler <dholler@redhat.com>