Prevent deleting of routers with ports
If a router is connected to ports, the router cannot be
deleted. If the router would be deleted, the connected lrp
would be removed along with the router, but the corresponding
lsp and dhcp-options (subnet) would still point to the
removed entities. This way the data base would be in an
inconsistent state.

Change-Id: Ied1f9f94d588a87faa7024dce2721afe350b867d
Signed-off-by: Dominik Holler <dholler@redhat.com>