Add requested-chassis to port
The patch adds the ability to set the port
binding:host_id attribute.
Binding the port to a specifis host id avoids
a race between hosts for the port during migration.
This port attribute is implemented by setting
the LSP options:requested-chassis value.
This will only be done if the PROVIDER.ovs-version-2.9
configuration value is set to true. This configuration
value indicates that the version of each ovn controller
in the system is at least 2.9. This must be done
because of BZ 1561499

Change-Id: Ib8b6c342ec4d37ea02a7dc36db89975ef443b3ae
Signed-off-by: Marcin Mirecki <mmirecki@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1549033