Raise generic ElementNotFoundError
If an OVN element is not found during the execution of an
ovsdbapp command, a generic ElementNotFoundError is raised,
which results in a HTTP 404 reply an OpenStack Networking API.
The error message will contain the related OVN tables and
names. On a higher level, the ElementNotFoundError can be
modified to create a nicer error message, if required.

Change-Id: Iebdfe649732270e9417759f8340cbf75d24114d3
Signed-off-by: Dominik Holler <dholler@redhat.com>