Fix the gerrit triggered CI
We install requests mock through pip, and the 1.5 version brought
along a new pytest_plugin that conflicts with the shipped version
of pytest.

Our shipped version of pytest asserts the following test:
  '''assert type in (None, "pathlist", "args", "linelist")'''
in the lib/python2.7/site-packages/_pytest/config.py, function
addini.

The pytest plugin added on requests-mock, version 1.5 features
the line:
  '''parser.addini('requests_mock_case_sensitive',
	'....',
	type='bool',
	default=False)'''

type=bool *is not* on the tuple that the shipped version of pytest
has, hence this failure.

It can also be seen in the jobs configs that the builds started to
fail when pip started to download version 1.5 of requests-mock.

Easiest way out is to fallback to version 1.4 of the requests-mock.

Change-Id: I7050c5e29dab30cc05a9897b604862dd10cc392f
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>