Add optional MTU parameter to network
When a created subnet is associated to a network that has an MTU
defined, we set the corresponding dhcp_options MTU to the value
defined in the network object.

This indirection is required because the MTU is an L2 parameter
on openstack networking API and an L3 paramenter (dhcp options)
on the ovn northbound API.

Thus, when a Network object in created, we store its MTU in the
external_ids column, and retrieve it when an associated subnet
object is created.

Change-Id: I116c4c7c49aa00639130fc97202a23618e5a7bf7
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>