automation: define stdciv2 jobs for el7 and fc27
According to the maintainers of ovirt CI, v2 is the way forward. Using
it allows ovirt-provider-ovn to choose which distros and branches should
be tested, without requiring a commit to the ovirt-jenkins project.

I find it harder to read the logs and build artifacts of v2 (they have
at least one extra level of nesting), but let us try to go with the
flow.

Change-Id: Ie0ab0dc4de5516ed09d18bcbc4bcbf432bc7f77e
Signed-off-by: Dan Kenigsberg <danken@redhat.com>