Limit the networks MTU to a maximum value
Add a configuration parameter where the maximum value of the MTU
can be set. If validation-max-allowed-mtu is kept at its default
value 0 zero, no validation is done.

When a user attempts to create/update a network specifying an MTU
larger than the configured max, an error is thrown.

If not set a user would succeed to set whatever kind of mtu he so
desires, which might be too big for his cluster connectivity, 
leaving the network behavior unpredictable. An ovirt administrator
may find it useful to protect his users from this condition.

This feature is also helpful when the ovn network is attached to a
physical network - the administrator could set this value to the
underlying network MTU, thus helping the user reach an expected
network throughtput.

Lastly, this feature mimics the neutron ML2 plugin behavior.

Change-Id: Ia56d1d4e9cca0f1f3e1511e9fa3c8ccad50668d9
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>