Allow duplicate network names
Store the network name (networking API) in the OVN logical switch
external ids column. We do not store the network name as plain
logical switch name because the logical switch name has to be
unique in OVN.

When mapping the responses from the OVN domain to REST, we extract
the name from the external ids column - if possible. If not possible
we return the actual name.

This last item is done to achieve backwards-compatibility.

Change-Id: I2ea78b1b16c5d0483099c5d7ce7a6581d014f36a
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1503577