Support vhostuser in vdsm hooks.
This patch enables vhostuser support via the driver. For OVS netdev
environments, instead of modifying the network interfaces to connect
to a regular OVS bridge (by using the virtualport attribute), this
patch creates a vhostuser server (qemu)/client (ovs) pair.

The server represents a network interface inside the guest, and the
client represents the OVN port and is connected to a localnet for
outside communication, i.e.:

eth0 <--> vhostuser server (qemu) <--> vhostuser client (ovs) <-->
br-int <--> localnet <--> outside

Until https://bugzilla.redhat.com/1598435 is resolved, selinux
has to be either disabled or set to permissive for qemu to be able
to create the vhostuser socket.

Change-Id: I7d0e7d9c9723e18ead9611e8ad1cd61f50998214
Signed-off-by: Leon Goldberg <lgoldber@redhat.com>