Eliminate redundant transactions
The _get_subnet_gateway_router_id method had as a parameter the
subnet_id, and used the ovsdbapp client to fetch the ovn dhcp
options object.

Since all the callers of this method already had the subnet object
available to them, that transaction is redundant - and by passing
the subnet object as a parameter it can be safely removed from the
code.

Change-Id: I98e2296c85fea8d4e3eb7eeeff90006d745028f2
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>