Implement basic query filtering
The patch adds basic query filtering for GET requests that
list entities.
The filter will filter all GET requests which return a list
of one type of an entity, for example:
  {"ports": [ {<port>}, {port}]}
The filtering will be done on the attributes of the entities
in the list.
For example if the query is:
  ports/mac_address=1
the result will include only the ports whose mac_address is 1.
If the filtered entity does not have an attribute matching the
filter key, it will not be included in the results.

Change-Id: I3c920002ffa3b80920a883933609ebd435418bb2
Signed-off-by: Marcin Mirecki <mmirecki@redhat.com>