Implement default sec group
Create a default port group - when port security is activated for
any port - featuring the following set of rules:
  - deny-all ingress traffic
  - deny-all egress traffic
  - allow ipv4 egress traffic
  - allow ipv6 egress traffic

The deny-all rules are not meant to be listed to the user,
and the user is not able to delete them. They implement the
security group rule behavior of dropping all traffic by default.

The allow ip4/ip6 egress traffic is meant to mimic openstack
networking API behavior. Those rules are listed to the user, and
can be removed using the API, should the user desire.

Change-Id: Ice7648b5601a566d87eb458d6e80f7611257eebc
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>