Implement the security group attibute in ports
We store the list of security groups assigned to each port
on the port external ids for performance reasons.

An alternative would be to filter all the port_group ovn
entities having the port id in its port list - which would take
unreasonable amounts of time on a simple 'list_ports' operation -
obviously, depending on the number of ports.

The information stored in the external_ids could be removed anytime
later, by parsing the port-groups and getting that info from there.
Therefore we do not create any backwards-compatibility trap (like
for example with network/port names stored in external_ids).

Change-Id: I044d022d98129cefc06704f13da8f97b48be2074
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>