Preserve compile timestamp of py source files
The modification timestamp of the py source code file is
embedded in the pyc bytecode file, which can be extracted by:
od --skip-bytes=4 --read-bytes=4 --address-radix=n --format=u4

If the timestamp embedded in the pyc file is older than the
modification time stamp of the py file, the python interpreter
will regenerate the pyc file at runtime.
To prevent the regeneration of the pyc files, the timestamp of
the py source files has to be preserved during install.

The regeneration of the pyc files after package install would
break 'rpm -v'.

Change-Id: Iffcfe58953631ed1f4e5be458d74335c9fdf867d
Bug-Url: https://bugzilla.redhat.com/1617985
Signed-off-by: Dominik Holler <dholler@redhat.com>