Keep track of security group IP addresses.
The remote_group_id paramerater implementation relies
on address sets. Using it, we keep track of all the IPs
in a security group.

When a rule is created, featuring the remote_group_id parameter,
we match the src/dst IP on the packet against the address set
identified by the remote_group_id parameter *and* the IP version.

Change-Id: I131b573e04252ceaacafe6db7dcec552270daa3e
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>