Fix: The ACL directions were upside down
In OVN, the direction concept is centered around the
logical port - e.g. traffic originating *from* the VM, is
using the 'from-lport' direction, whereas traffic being
sent *to* the VM uses the 'to-lport' direction.

Furthermore, OVN features 2 distinct pipelines - ingress,
and egress. ACLs having the 'to-port' direction are applied
in the *egress* pipeline, and ACLs having the 'from-lport'
direction are applied in the *ingress* pipeline.

In the end, the networking API security group rules with
and *egress* direction have to be written to the ovn
*ingress* pipeline, whereas rules with the *ingress*
direction have to be written to the *egress* pipeline.

Since the test I had performed only accounted for symmetric
traffic - e.g. both the src and dst VM belonged to the same
group - this bug was not found.

Change-Id: I4025717cda6a21ee5f60961d43fb8d06672fa820
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>