Tox based integration test skeleton
This patch adds another tox environment, meant for integration
testing.

The requirements, scope, and plan for this integration test
environment can be found in [0].

The integration tests can be executed in 2 ways:
  - tox -e it-tests27
  - make integrationtest

This new test environment is meant to be used in the check-patch
and check-merged CI stages - and manually, of course.

This patches only updates the check-merged script because the check-
patch script is a symlink to the check-merged one.

[0] - https://gerrit.ovirt.org/#/c/94562/

Change-Id: I26a7a846e69d54731107b70b5f6f5e09e074a277
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>