Create provider container.
This patch finishes building the test environment, by pulling and
configuring the ovirt_provider_ovn container.

The source code of the git repo HEAD is mounted on the container,
which is downloaded from dockerhub. The RPMs are afterwards
generated, and installed in the ovirt-provider-ovn container.
Finally, the ovirt-provider-ovn service is started.

Port forwarding is configured in the container for the neutron and
keystone ports - all localhost traffic intended for ports 9696 and
35357 is redirected into the container.

It also updates the dummy integration test to query the provider
for a network created out of band - in the create_it_env.sh script.

Change-Id: Iee79b8e0ba9449fc67245acb211d98709c12b656
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>