Configurable ovn dependencies
The ovn containers being pulled from triple-O project were tagged
currenttriple-o.

This patch turns that into the default, so the user can override
the tags if desired.

Mapping from dockerhub tags to release / git commit is a problem
of its own, and overriding the tags should be seen as not advised.

Change-Id: I2a99c402ea23dd3cb9b416f4e4c5f4d2689a85af
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>