integration-tests: test security groups api
Brought the ansiblelib module from the OST repo, along with the
callback plugins defined in it. The sources can be found in [0],
and [1].

The ansible playbook was copied from the openstacksdk project,
whose source is in [2].

The ansiblelib tests idempotency of the playbooks, by executing
them twice - this lead to the single reusable role defined at
[2] to be split into 2 different playbooks - one for create,
another for delete entities.

Due to a current bug in the os_security_group_rule module- [3],
the delete rule operation is not idempotent, thus the cleanup
temporarily will delete only the group, which, during deletion
will delete the rules.

A final remark, is that the ansiblelib looted from OST was
adapted, allowing the idempotency check to be optional. The default
behavior is to perform the idempotency check. This is useful since
some playbooks might model a procedure, instead of a desired state.

[0] - https://github.com/oVirt/ovirt-system-tests/blob/master/network-suite-master/lib/ansiblelib.py
[1] - https://github.com/oVirt/ovirt-system-tests/tree/master/network-suite-master/ansible/callback_plugins
[2] - https://github.com/openstack/openstacksdk/blob/master/openstack/tests/ansible/roles/security_group/tasks/main.yml
[3] - https://github.com/ansible/ansible/issues/50057

Change-Id: I4dab4b0dbf93210b53960276bbef9f04a6bf2125
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>