integration-tests: create_l2l3_scenario role
Introduces the 'network_point' concept, which represents a network
client in the dataplane.

A network point represents a logical port, but it also features
all L2 and L3 information for the port's networking API dependencies
, which are networks, and subnets.

All the network_points in the dataplane are iterated, creating the
respective elements; first all the networks are created, then all
the subnets, and finally, the ports are created.

This concept can be used to create multiple namespaces in the same
network and subnet, or multiple namespaces in different networks,
having different subnets.

This patch is the first of a set of patches intended to create a
reusable way of creating dataplane configurations, which will be
leveraged by the integration tests. The other patches are:
  - https://gerrit.ovirt.org/#/c/96170
  - https://gerrit.ovirt.org/#/c/96171
  - https://gerrit.ovirt.org/#/c/96172
  - https://gerrit.ovirt.org/#/c/96173

Change-Id: I396aeb7cb810a999b83ec3cf15f56bd485a7c5b9
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>