integration-tests: create a reusable scenario
This commit leverages the roles introduces in previous
integration-tests patches to generate a data plane scenario
from control plane entities.

The playbook requires the id of the ovn-controller container
and the id of the ovn-central container, adding those to the
inventory.

Also, it requires a list of 'network_points' to create the
control plane (ovn entities), and setup and configure the
OVS ports.

Finally, if the 'ping_destination_ip' and 'ping_source_namespace'
are defined, the playbook tests the dataplane it set up through
a ping, from one IP namespace to the desired IP.

Change-Id: Ica843db5a8f23faa1df8ae36bfb15b11e8933ed4
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>