fix: disable port security was broken
Currently, when port security is disabled for a port, we remove
the association of that port to the 'Default' port group.

This used to mean that traffic would be resumed, since the Default
port group had the 'deny' OVS flows, until commit 9fead7f9c5.

In that commit, the 'deny' flows were moved to a different, deny
specific port group, and the name of the group where the port
association was severed was not updated.

This commit updates the name of the group where we delete the
port association when port security is disabled, and also removes
some duplicated related code.

Change-Id: Ic924f108508d5f4d76a51ecd2ae4d5a7ca5fdf1b
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>