integrate w/ manageIQ on ovirt 4.2
Currently, the security groups API attempts to query through
ovsdbapp the ovn-central instance. In ovirt-4.2, the versions of
ovsdbapp & ovn central do *not* feature the Port_Group table, which
makes the provider return a 500 status code to manageIQ.

This patch changes that, by querying the ovn-central schema, and
checking if the required table is available.

If the required table is not available, we return a 405 status code
which would comply with the previous provider behavior (e.g. when
the provider didn't have security groups implemented).

The "GET security-groups" API method is different, since manageIQ
requires a list to be read, according to commit
b0b781a2413b4ca695f1bf6002b2f87fdc82d1d6.

Checking if the PortGroup table is available on the schema
is the prefered way to check this, since the schema is dynamically
loaded from the ovn-central instance the provider is connected to.

Addresses issue: https://github.com/ManageIQ/manageiq/issues/18457

Change-Id: Ifeb8f793a387722289d7a4cc2788fc4d2120b7ea
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>