Pre-empt provider log file creation.
Also, separate ovn/provider data retrieval process.

The ovn logs and DBs, along with the provider-ovn logs are
stored away in the exported-artifacts directory after the
integration tests are executed. This happens on success *and*
error scenarios. It is extremelly useful for CI post-mortems.

The script saves the provider logs as long as the provider
container is up and running - it makes no check to see if the
log file actually exists. Since this was done in the environment
teardown function, the containers would be left behind in the
environment on any errors in the teardown function.

By separating the logs retrieval function, and by creating the
log file before attempting to create the provider rpm we can avoid
cascading error scenarios, thus assuring the containers are removed
at the end of the test execution.

Change-Id: I455fb85085ab804fac91e5822a25f56cd1df1880
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>