Stateless dhcpv6 does not support fixed IPs
When fixed IPs are used to create an ovn logical port, dhcp
logical flows are provisioned to make the OVN dhcp service
reply with the pre-provisioned IP addres.

When the subnet is configured with address mode stateless,
the OVN native DHCP server *does not* add the IP address to
the DHCPv6 reply message, which renders the feature unusable.

Thus, fixed IPs are not supported when the port is created
in a subnet that is configured to opperate statelessly.

Change-Id: I72b6c28ed8d3bcbd6b18f556afccd5183b605a81
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>