ipv6: add mtu option to RA messages
To configure the MTU for networks having an ipv6 subnet, the
network MTU parameter has to be used to configure the logical
router port ipv6_ra_config.

Thus, the network's MTU is stored within the subnet's external
ids, and when a router interface is added, the MTU is extracted
from the subnet and used to configure the logical router port.

For network MTU updates, a similar approach is taken; all logical
router ports associated to the subnet attached to the network are
filtered, and those ports ipv6_ra_config is updated.

To prevent network misconfiguration, a constraint forcing the
gateways per router to be unique was added.

Also introduce an integration test to assert the MTU configured
via RA messages:
  - with default MTU (1442) pinging with a data size of 1301
    succeeds
  - one of the 2 networks MTU is updated to a smaller value (1300)
  - pinging with a ping data size of 1301 fails

To perform the MTU update in the guests (IP namespaces), a
function to reconfigure the interfaces was added, where the
interface is disabled, then enabled; this leads to an RA
message being sent, to which the OVN logical router replies with
an RA message, featuring the new MTU. With that message, the
interface can be successfully configured with the new MTU.

Bug-Url: https://bugzilla.redhat.com/1685983
Change-Id: Ibca14e0f26a4fcf885435789684868f44dd01703
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>