integration-tests: configure the ovn-controller in the boot script
New versions of ovn-controller require the ovn configuration to
be set before starting up the controller, otherwise, a default
encapsulation is created, which in turn will trigger a constraint
error when attempting to configure it from within the
create_it_env.sh script.

Also set the controller's logging to debug.

Change-Id: Ifd6535803100e5d2c0cb3b1f34b10703d6ba22d5
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>