integration-tests: dataplane debug tools
Optionally start tcpdumps on the dataplane guests to filter
icmp6 or dhcpv6 traffic to debug the guest IP addressing
configuration.

This is helpful both for debugging, and to provide out of
band captures for OVN team - so far that is what I used it
for.

The current code only supports capturing icmp6 / dhcpv6 traffic
for *a single* test, since the captures will be overwritten.
Follow up patches will extend this behavior, enabling it to
be used on multiple tests, and to extend it to other kinds of
traffic.

To activate this feature for one test, the user should:

1. start the test environment
  > automation/create_it_env.sh

2. activate debugging on the test input
   ...
    'provider_container_id': PROVIDER_CONTAINER_ID,
    'controller_container_id': CONTROLLER_CONTAINER_ID,
    'debug_icmpv6': True
  }

3. run the test
  > tox -e integration-tests27 -- \
      test_ipv6_subnets.py::test_multiple_subnets

4. copy the captures from the container
  > docker cp <ovn_controller_container>:/root/lport1.pcap <dst>
  > docker cp <ovn_controller_container>:/root/lport2.pcap <dst>

Change-Id: I7c89996d22229df709452eeb70ad1e44ca2a21fb
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>