fix error handling for python3
On python3 the send_error function sets the content lenght
header based on the size of the message string passed to it
as argument.

Our code is not using that function to send the error, and
instead writes it directly via the server stream write function.

This change keeps the py2 behavior, while correctly uses the
send_error function on python3.

Change-Id: I5a6214029f81f2c6a86765447d51f997165935d4
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>