transactions: set the port name as a generated id
Also updated the ovn_north.get_lsp function to *always*
return a logical switch port, it made no sense for it
to sometimes return an LSP, other the ID of an LSP; it
was also required, so we can read the lsp from the
lrp_id, then extract the lrp *name* from it.

The lsp_get command now allows for the client to query lsps
by lsp_name, which does exactly what get lsp by id does.

Change-Id: I19071271a9ce14a209f62689991e9dbaa12190fc
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>