transactions: add a method to build an update command
The element to add to any ovsdbapp transaction is a Command, and
as such, the logic to create an update command had to be separated
from the 'execute' method.

Now the 'execute' method invokes 'create_command', then 'executes'
while it also allows for building commands to be added to ovsdbapp
transactions.

Change-Id: I4d0503ea357a06290916a3746fda66ccfc693dda
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>