transactions: optionally commit Commands within transactions
Add a decorator that will add the respective command to a
transaction, if one is passed as an argument to the command
builder.

When a transaction is not added, the decorator will execute the
command directly.

This way, moving the provider code to the performed within
transactions can be done iterativelly.

Change-Id: Ib1a33bc7363bbe21ec0a735fea87a7dc9b8529f9
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>