Move security group related constants
This commit prepares the ground for the security group / rules
transaction patches, by moving the security group related
constants to the SecurityGroupMapper class.

Since creating the security groups within transaction require
an update to the security group / security group rule association
(associate via name, other than via id), to keep matters backwards
compatible, a new key will be introduced higher in the patch stack.

This key will indicate if the mapping between rules should be
performed by ID or by name.

In the neutron_api_mappers, we'll use this key to know which
id (actual ovsdb row uuid vs ovsdb row name) to return. For the
DropAll and Default group we need to always return the uuid. This
change will also be performed in a separate patch.

This last requirement means that the SecurityGroupMapper layer
requires awareness of the group names - to exclude the
'whitelisted' group names.

Change-Id: I7d1123a398787b621fdbd000f7cde10fed27adcc
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>