neutron api: return the correct group ID
When the provider did not use transactions, the association
between security groups and rules was done by ID - e.g. storing
the security group ID in one key in the security group rules
external IDs.

Transaction abilities will be introduced in the provider project,
and this association will have to performed by 'name' - the UUID of
the group is *not* available during rule creation.

This patch updates the security group mappers to return the correct
security group ID - either 'name' or 'uuid'.

Change-Id: Iae23d12485ccc4f63d248018d4259c5a78ab5487
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>