transactions: add default sec GID wrapper decorator
To preserve backwards compatibility, we need to reply with the
group uuid in the 'security_group_id' field of the security group
rules.

To keep the code impact minimal, we will rely on this decorator,
which will pass to the mapper layer the security group rule object
wrapped within a class that has a static method that indicates
which is the default group id.

Change-Id: Id39e4670fa0c39d43bc881f82dcca6d0f3f8e443
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>