Only activate/deactive the default sec group when required
Up until this patch, we were assigning a the ports that had
port_security_enabled to the Default / DropAll security groups
simply based on the current value of port security.

With this patch, we only do so when the state changes - from
disabled to enabled, we associate the port to the corresponding
port groups; from disabled to enabled, we remove the port
association.

The old behavior was causing the provider code to fail on 4.2,
where the OVS version used does not feature the port group tables,
since we were trying to dissociate the port from the port group
whenever the port_security_enabled=False.

Change-Id: I9b4d8dc816f0084ad9176248d85e5de88f1a1bc6
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>