Fix deploy failure on fedora hosts
The current spec was installing the vdsm-tool ovn configuration
script in python3 sitelib, while vdsm-tool is being run on python2.

Until vdsm is ported to py3, and vdsm-tool runs under py3 - which
has to occur for fedora30, we need to use python2 on the driver.

The fedora image - in which the integration-tests3 target is executed
against - is updated to figure fedora30, thus enabling python3
interpreters.

Change-Id: I5e9e4f7d4eb4a9fccf3dd5ca87557df0b4dd9364
Bug-Url: https://bugzilla.redhat.com/1725013
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>