Let ovn-controller keep track of port addresses
We are currently manually keeping track of the port's IP addresses
in the ovn north table address sets. For this reason, and since
dynamic addresses are assigned by the ovn controller automatically,
we cannot create a logical port in a single transaction.

The ovn controller itself keeps track os the port's IP addresses in
the ovn *south* database, so all code interacting w/ the address set
tables can be safely removed.

Change-Id: I419d7cb04ef3c6725e8d3444fe34e3c27773ea32
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>