Create network within transaction
The current network name is not updated, since we already enforce
unique network names - by generating a UUID and use it as part of
the name string.

Add an integration-test that assures the correct creation of a
physnet via the provider.

The OVN interaction code was updated to directly use the localport
'tag' parameter instead of the 'tag_request' parameter. Using the
'tag_request' was keeping the door open for a race condition, since
it was ovn-north that copied the contents of 'tag_request' and wrote
it into 'tag' - which is the parameter from which we report the
segmentation_id networking API parameter.
Since this out done out of the scope of our transaction, it would
sometimes take too long, and we would report no VLAN.

Change-Id: I92463f87193ad84f740cfe6c951b0f35cc63643a
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>