integration-tests: expand on port security test
Update the security group tests, having one where when port
security is activated for all ports, and all ports are added
to the default port group; and another test where port security
is activated, but only *one* of the ports is assigned to the
default group.

In the former test connectivity between the namespaces is expected,
but in the latter test connectivity should not exist, since one of
the ports is not present in the Default group - where intra-group
connectivity exists.

Change-Id: I01fdfba2ef6d03695b08901ec7f94ca4ac2dc06a
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>